requirement.txt
mysqlclient
django-wysiwyg
django-tinymce
djangorestframework
markdown
django-filter
